Franco-Iberian War Events
Finished some localisations, and added events for the scenario where Aragon decides to keep Naples. In this case, rather than a war against the Italian League, the first Italian War that triggers is a Franco-Iberian war.

Instead of the normal event to choose if they want to join a side, the Italic League won't be able to intervene until one side is at -25 war score. Then they can join the losing side, declare on both, or not intervene, in which case the league is passed to someone else and they intervene (this option is only included for the player's freedom). And I adjusted IW_flavor.1 so that, in this scenario, the Imperial League will then form and join the Italic League's offensive wars to defend Imperial lands if the Italians start to lose